Add per connection IO statistics and update the adhoc browser to make it possible to query them.

+++ src/frontend/org/voltdb/compiler/AsyncCompilerResult.java
@@ -24,7 +24,7 @@ public class AsyncCompilerResult implements Serializable {
 
     public long clientHandle = -1;
     public String errorMsg = null;
-    public int connectionId = -1;
+    public long connectionId = -1;
     public String hostname = "";
     transient public Object clientData = null;
